- Refactore service and dao layer classes using coding to interface principle - Add javadoc documentation
