+++ src/main/java/com/unitedremote/bootcamp/dao/UserDaoImpl.java
@@ -9,19 +9,22 @@ import com.unitedremote.bootcamp.models.User;
 import com.unitedremote.bootcamp.repositories.UserRepository;
 
 @Service
-public class UserDao {
+public class UserDaoImpl  implements UsuerDao{
 	
 	@Autowired
 	UserRepository userRepository;
 	
+	@Override
 	public List<User> findAll() {
 		return userRepository.findAll();
 	}
 	
+	@Override
 	public User findOne(Long id) {
 		return userRepository.getOne(id);
 	}
 	
+	@Override
 	public void delete(User user) {
 		userRepository.delete(user);
 	}
