+++ src/main/java/com/unitedremote/bootcamp/services/SkheraService.java
@@ -3,63 +3,44 @@ package com.unitedremote.bootcamp.services;
 import java.util.Date;
 import java.util.List;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.unitedremote.bootcamp.dao.SkheraDao;
 import com.unitedremote.bootcamp.models.Rider;
 import com.unitedremote.bootcamp.models.Skhera;
 
-@Service
-public class SkheraService {
+public interface SkheraService {
+
+	Skhera saveSkhera(Skhera skhera);
 	
-	@Autowired
-	SkheraDao skheraDao;
+	Skhera findOne(Long id);
 	
-	private static Logger logger = LogManager.getLogger();
-	
-	public List<Skhera> findAll() {
-		return skheraDao.findAll();
-	}
-	
-	public Skhera findOne(Long id) {
-		logger.debug("finding Skhera with id: " + id);
-		Skhera skhera= new Skhera();
-		try {
-			skhera = skheraDao.findOne(id).get();
-		} catch (Exception e) {
-			logger.warn("Error: " + e.getMessage());
-		}
-		return skhera;
-	}
-	
-	public Skhera saveSkhera(Skhera skhera){
-		logger.debug("saving Skhera ...");
-		return skheraDao.save(skhera);
-	}
-	
-	public void deleteSkhera(Skhera skhera) {
-		logger.debug("deleting Skhera with id: " + skhera.getId());
-		skheraDao.delete(skhera);
-	}
-	
-	public Skhera setRider(Skhera skhera, Rider rider){
-		logger.debug("setting rider " + rider.getId() + "to skhera "  +skhera.getId());
-		 skhera.setRider(rider);
-		 return this.saveSkhera(skhera);
-	}
-	
-	public Skhera pickupSkhera(Skhera skhera, Date date) {
-		logger.debug("picking up Skhera " + skhera.getId() + " at "+date.toString());
-		skhera.setPickedUpAt(date);
-		return skheraDao.save(skhera);	
-	}
-	
-	public Skhera deliverSkhera(Skhera skhera, Date date) {
-		logger.debug("delivering Skhera " + skhera.getId() + " at "+date.toString());
-		skhera.setDeliveredAt(date);
-		return skheraDao.save(skhera);	
-	}
+	List<Skhera> findAll();
+
+	void deleteSkhera(Skhera skhera);
+
+	/**
+	 * Assigns a rider to a skhera
+	 * 
+	 * @param skhera
+	 * @param rider
+	 * @return skhera
+	 */
+	Skhera setRider(Skhera skhera, Rider rider);
+
+	/**
+	 * Add a date value identifying the skhera picking up by the rider assigned
+	 * 
+	 * @param skhera
+	 * @param date
+	 * @return skhera
+	 */
+	Skhera pickupSkhera(Skhera skhera, Date date);
+
+	/**
+	 * Add a date value identifying the skhera delivering by the rider assigned
+	 * 
+	 * @param skhera
+	 * @param date
+	 * @return skhera
+	 */
+	Skhera deliverSkhera(Skhera skhera, Date date);
+
 }
