+++ src/main/java/com/unitedremote/bootcamp/services/utils/UnableToShareSkherasRidersComparator.java
@@ -4,6 +4,11 @@ import java.util.Comparator;
 
 import com.unitedremote.bootcamp.models.RiderComparable;
 
+/**
+ * According to this criteria: one comparable rider is "better" than another,
+ * if he can get to the pickup address of a given skhera in less time
+ *
+ */
 public class UnableToShareSkherasRidersComparator implements Comparator<RiderComparable> {
 
 	@Override
